feat: 받은 봉투 생성 기능 구현

+++ core/model/src/main/java/com/susu/core/model/Friend.kt
@@ -1,8 +1,8 @@
 package com.susu.core.model
 
 data class Friend(
-    val id: Int = 0,
-    val uid: Int = 0,
+    val id: Long = 0,
+    val uid: Long = 0,
     val name: String = "",
     val phoneNumber: String = "",
     val createdAt: String = "",
